+++ src/models/svc.py
@@ -19,7 +19,7 @@ class SVEstimator():
         model_params = {k: v for k,v in params.items() if k not in ["cv_folds","beta","scaler","pca","model","C_min","C_max","n_trials","pca_level","scaler_type"]}
         self.classifier = SVC(**model_params)
 
-    def cross_validate(self, X_train: np.array, y_train: np.array) -> np.float:
+    def cross_validate(self, X_train: np.array, y_train: np.array) -> float:
         splitter = StratifiedKFold(n_splits=self.cv_folds, shuffle=True, random_state=self.classifier.random_state)
         score = 0.0
         self.cv_scores = {
